+++ fleetfinder/tests/test_templatetags.py
@@ -10,6 +10,7 @@ from django.test import TestCase, override_settings
 from fleetfinder import __version__
 from fleetfinder.constants import PACKAGE_NAME
 from fleetfinder.helper.static_files import calculate_integrity_hash
+from fleetfinder.templatetags.fleetfinder import get_item
 
 
 class TestVersionedStatic(TestCase):
@@ -93,3 +94,61 @@ class TestVersionedStatic(TestCase):
 
         with self.assertRaises(ValueError):
             template_to_render.render(context=context)
+
+
+class TestGetItem(TestCase):
+    """
+    Test the `get_item` template tag
+    """
+
+    def test_returns_value_for_existing_key(self):
+        """
+        Test should return the value for an existing key
+
+        :return:
+        :rtype:
+        """
+
+        dictionary = {"key1": "value1", "key2": "value2"}
+        result = get_item(dictionary, "key1")
+
+        self.assertEqual(result, "value1")
+
+    def test_returns_none_for_non_existing_key(self):
+        """
+        Test should return None for a non-existing key
+
+        :return:
+        :rtype:
+        """
+
+        dictionary = {"key1": "value1", "key2": "value2"}
+        result = get_item(dictionary, "key3")
+
+        self.assertIsNone(result)
+
+    def test_returns_none_for_empty_dictionary(self):
+        """
+        Test should return None for an empty dictionary
+
+        :return:
+        :rtype:
+        """
+
+        dictionary = {}
+        result = get_item(dictionary, "key1")
+
+        self.assertIsNone(result)
+
+    def test_returns_none_for_none_dictionary(self):
+        """
+        Test should return None for a None dictionary
+
+        :return:
+        :rtype:
+        """
+
+        dictionary = None
+        result = get_item(dictionary, "key1")
+
+        self.assertIsNone(result)